+++ libraries/docker_image.rb
@@ -16,11 +16,20 @@ module DockerCookbook
     property :rm, [true, false], default: true
     property :source, String
     property :tag, String, default: 'latest'
+    property :buildargs, [String, Hash], coerce: proc { |v| v.is_a?(String) ? v : coerce_buildargs(v) }
 
     alias_method :image, :repo
     alias_method :image_name, :repo
     alias_method :no_cache, :nocache
     alias_method :no_prune, :noprune
+
+    ###################
+    # Property helpers
+    ###################
+
+    def coerce_buildargs(v)
+      "{ #{v.map { |key, value| "\"#{key}\": \"#{value}\"" }.join(', ')} }"
+    end
 
     #########
     # Actions
@@ -93,6 +102,7 @@ module DockerCookbook
           {
             'nocache' => new_resource.nocache,
             'rm' => new_resource.rm,
+            'buildargs' => new_resource.buildargs,
           },
           connection
         )
@@ -105,6 +115,7 @@ module DockerCookbook
           {
             'nocache' => new_resource.nocache,
             'rm' => new_resource.rm,
+            'buildargs' => new_resource.buildargs,
           },
           connection
         )
@@ -117,6 +128,7 @@ module DockerCookbook
           {
             'nocache' => new_resource.nocache,
             'rm' => new_resource.rm,
+            'buildargs' => new_resource.buildargs,
           },
           connection
         )